Move Guild related files to new Guild directory
