+++ Source Main 5.2/source/GMBattleCastle.cpp
@@ -21,7 +21,7 @@
 #include "GIPetManager.h"
 #include "BoneManager.h"
 
-#include "GuildCache.h"
+#include "Guild/GuildCache.h"
 #include "ZzzInterface.h"
