+++ Source Main 5.2/source/Guild/NewUIGuildMakeWindow.h
