+++ Source Main 5.2/source/Guild/UIGuildMaster.h
