+++ Source Main 5.2/source/NewUICommonMessageBox.cpp
@@ -3,13 +3,13 @@
 #include "stdafx.h"
 #include "NewUICommonMessageBox.h"
 #include "NewUICustomMessageBox.h"
-#include "NewUIGuildMakeWindow.h"
-#include "NewUIGuildInfoWindow.h"
+#include "Guild/NewUIGuildMakeWindow.h"
+#include "Guild/NewUIGuildInfoWindow.h"
 #include "NewUIMyInventory.h"
 #include "UIControls.h"
 #include "ZzzBMD.h"
 #include "ZzzCharacter.h"
-#include "UIGuildInfo.h"
+#include "Guild/UIGuildInfo.h"
 #include "UIManager.h"
 #include "PersonalShopTitleImp.h"
 #include "CComGem.h"
