+++ Source Main 5.2/source/NewUIGoldBowmanLena.h
@@ -6,7 +6,7 @@
 #include "NewUIInventoryCtrl.h"
 #include "NewUIMessageBox.h"
 #include "NewUIMyInventory.h"
-#include "NewUIGuildMakeWindow.h"
+#include "Guild/NewUIGuildMakeWindow.h"
 
 namespace SEASON3B
 {
