+++ Source Main 5.2/source/NewUIMiniMap.cpp
@@ -9,7 +9,7 @@
 #include "NewUICustomMessageBox.h"
 #include "DSPlaySound.h"
 
-#include "NewUIGuildInfoWindow.h"
+#include "Guild/NewUIGuildInfoWindow.h"
 #include "NewUIButton.h"
 #include "NewUIMyInventory.h"
 #include "CSItemOption.h"
