+++ Source Main 5.2/source/NewUISystem.h
@@ -8,7 +8,7 @@
 #include "NewUIHotKey.h"
 #include "NewUIChatLogWindow.h"
 #include "NewUISlideWindow.h"
-#include "NewUIGuildMakeWindow.h"
+#include "Guild/NewUIGuildMakeWindow.h"
 #include "NewUIFriendWindow.h"
 #include "NewUIMainFrameWindow.h"
 #include "NewUIChatInputBox.h"
@@ -23,7 +23,7 @@
 #include "NewUIGateSwitchWindow.h"
 #include "NewUIStorageInventory.h"
 #include "NewUIStorageInventoryExt.h"
-#include "NewUIGuildInfoWindow.h"
+#include "Guild/NewUIGuildInfoWindow.h"
 #include "NewUIMyShopInventory.h"
 #include "NewUIPurchaseShopInventory.h"
 #include "NewUICharacterInfoWindow.h"
