+++ Source Main 5.2/source/WSclient.cpp
@@ -1,7 +1,7 @@
 ﻿#include "stdafx.h"
 #include <memory>
 #include "UIManager.h"
-#include "GuildCache.h"
+#include "Guild/GuildCache.h"
 #include "ZzzBMD.h"
 #include "ZzzInfomation.h"
 #include "ZzzObject.h"
