+++ Source Main 5.2/source/ZzzScene.cpp
@@ -3,7 +3,7 @@
 
 #include "stdafx.h"
 #include "UIManager.h"
-#include "GuildCache.h"
+#include "Guild/GuildCache.h"
 #include "ZzzOpenglUtil.h"
 #include "ZzzBMD.h"
 #include "ZzzInfomation.h"
